pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi",
    "uvicorn",
    "aiogram",
    "motor",
    "pydantic-settings",
    "google-genai",
]
